Switch HeadersValue and MetadataValue to use Double
For consistency with JSONValue, which seems like the right thing to do
API-wise and also will make some upcoming work easier.

+++ Sources/AblyChat/Headers.swift
@@ -2,7 +2,7 @@
 
 public enum HeadersValue: Sendable, Codable, Equatable {
     case string(String)
-    case number(Int) // Changed from NSNumber to Int to conform to Codable. Address in linked issue above.
+    case number(Double) // Changed from NSNumber to Double to conform to Codable. Address in linked issue above.
     case bool(Bool)
     case null
 }
